ECHAM always returns xarray objects; fixes list requirements (#5)
ECHAM analysis did not consistently return xarray objects. (never was an issue, but would have showed up sooner or later -- this raises the need for tests!)

The file list is now always computed inside of the function; so that the interface simplifies to only get a variable name. Probably, I need to figure out a way to create an abstract method which fails if you try to create a method that takes more than just a filename...